back/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic_settings
email-validator
pymongo
python-decouple
passlib
python-jose
python-multipart
bcrypt
faker
tqdm
pytest
httpx
fastapi-mail
apscheduler
isort
black
fastapi-pagination
python-slugify